+++ spec/vcr/http_stubbing_adapters/faraday_spec.rb
@@ -7,9 +7,9 @@ describe VCR::HttpStubbingAdapters::Faraday, :without_monkey_patches => :all do
     :status_message_not_exposed, :does_not_support_rotating_responses
 
   it_performs('version checking',
-    :valid    => %w[ 0.5.3 0.5.10 ],
-    :too_low  => %w[ 0.5.2 0.4.99 ],
-    :too_high => %w[ 0.6.0 1.0.0 ]
+    :valid    => %w[ 0.6.0 0.6.10 ],
+    :too_low  => %w[ 0.5.9 0.4.99 ],
+    :too_high => %w[ 0.7.0 1.0.0 ]
   ) do
     before(:each) { @orig_version = Faraday::VERSION }
     after(:each)  { Faraday::VERSION = @orig_version }